API: move rustc_serialize bindings to its own crate.
