+++ rmp-serialize/src/lib.rs
@@ -1,0 +1,8 @@
+extern crate rmp;
+extern crate rustc_serialize as serialize;
+
+pub mod decode;
+pub mod encode;
+
+pub use decode::Decoder;
+pub use encode::Encoder;
