+++ src/lib.rs
@@ -122,7 +122,6 @@
 //! ```
 
 extern crate byteorder;
-extern crate rustc_serialize as serialize;
 
 extern crate serde;
 
@@ -134,9 +133,6 @@ mod init;
 pub mod value;
 
 pub const MSGPACK_VERSION : u32 = 5;
-
-pub use decode::serialize::Decoder;
-pub use encode::serialize::Encoder;
 
 pub use decode::serde::Deserializer;
 pub use encode::serde::Serializer;
