+++ tests/func/encode/mod.rs
@@ -10,7 +10,6 @@ mod map;
 mod ext;
 
 // High-level serialize interface.
-mod encoder;
 mod serializer;
 mod value;
 mod value_ref;
