+++ tests/func/mod.rs
@@ -1,4 +1,3 @@
 mod encode;
 mod decode;
-mod rserialize;
 mod serde;
