+++ tests/lib.rs
@@ -2,7 +2,6 @@
 #![cfg_attr(feature = "serde_macros", plugin(serde_macros))]
 
 extern crate rmp as msgpack;
-extern crate rustc_serialize;
 extern crate serde;
 
 mod func;
